% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league-package.R
\docType{package}
\name{league-package}
\alias{league}
\alias{league-package}
\title{league: NBA Shiny League}
\description{
A Shiny application for NBA Fantasy
}
\author{
\strong{Maintainer}: Oliver Eaton \email{eat_fred@protonmail.com}

Authors:
\itemize{
  \item Oliver Eaton \email{eat_fred@protonmail.com}
}

}
